Clean up day 6.
The way to get this fast (part 1) is to compute for a given grid that is
just big enough, and then mark all regions containing the border points
as infinite. That way only one meshgrid has to be computed for, and it
just has to be a bounding box. Still, for really evil points
(sporadically placed) this could still lead to quite significant memory
issues. So I'll leave my generator version, despite vectorized versions
being faster.