finally nail day 16 2022
One of the worst ones to date, spent so long on a solution searching
through step by step, picking choices for open or move to neighbor each
time. Almost made it work for part 2 as well, reaching search depth of
23 minutes with some alpha pruning. Still, failed because it would
require >100GB RAM to run efficiently through it all... Got lost in
trying to do A* search with no improvement, because I had (incorrectly)
given up on DPS after an initial attempt.

The insights that made it work:

- Pre-compute the graph into pairwise connections with weighted edges
  giving the time it would take to get from one to the other.
- Depth first, memoized.
- In an optimal solution, the elephant doesn't try to open any of the
  valves that we open, so we can search the elephant moves on just the
  subset of valves that we haven't opened.

Total runtime: 1.3 seconds.